requirements.txt
certifi==2017.11.5
chardet==3.0.4
defusedxml==0.5.0
Django==1.11.7
django-allauth==0.34.0
django-contrib-comments==1.8.0
django-crispy-forms==1.7.0
django-filter==1.1.0
djangorestframework==3.7.3
djangorestframework-jwt==1.11.0
idna==2.6
Markdown==2.6.9
oauthlib==2.0.6
olefile==0.44
Pillow==4.3.0
PyJWT==1.5.3
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
urllib3==1.22
